Refactor Cars component to use a separate carsLoader and clean up imports
